Segregated interaction object system
