+++ Assets/Scripts/Objects/Interaction System Objects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 88df139326684cd45b33235f2d900668
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
