+++ Assets/Scripts/Objects/Interaction System Objects/DialogInteraction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 23ff8d003996e1345b84064bf7942aea
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
