+++ Assets/Scripts/Objects/Interaction System Objects/IInteraction.cs
