+++ Assets/Scripts/Objects/Interaction System Objects/InteractableObject.cs.meta
