+++ Assets/Scripts/Objects/Interaction System Objects/PickupInteraction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8b897ea57f9b349408a8aae20b1de89c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
