+++ Assets/Scripts/Objects/Interaction System Objects/PuzzleInteraction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 16da2b045d27c1e4fb12aa797e1da3cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
