+++ Assets/Scripts/Player/ActionController.cs
@@ -1,6 +1,5 @@
 using KrazyKrakenGames.DetectiveGame.Managers;
 using StarterAssets;
-using System.Linq;
 using UnityEngine;
 using static KrazyKrakenGames.DetectiveGame.Global.MetaConstants;
